test: move postMessage-return test into its own file
worker_threads.test.ts has pre-existing debug-build timeouts unrelated
to this change; putting the new coverage in a focused file lets the
gate verify fail-before/pass-after cleanly.
